requirements.txt
cffi==1.14.5
cryptography==3.4.7
Django==3.2.2
django-crispy-forms==1.11.2
django-filter==2.4.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
pycparser==2.20
pytz==2021.1
sqlparse==0.4.1